+++ plugins/src/main/kotlin/com/project/starter/modules/plugins/ConfigurationPlugin.kt
@@ -4,6 +4,7 @@ import com.project.starter.config.plugins.rootConfig
 import com.project.starter.modules.internal.configureCommonDependencies
 import com.project.starter.modules.internal.configureKapt
 import com.project.starter.modules.internal.configureRepositories
+import com.project.starter.quality.plugins.QualityPlugin
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
@@ -11,6 +12,9 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 internal class ConfigurationPlugin : Plugin<Project> {
 
     override fun apply(target: Project): Unit = with(target) {
+        if (rootConfig.quality.enabled) {
+            pluginManager.apply(QualityPlugin::class.java)
+        }
         configureKapt()
         configureRepositories()