+++ Changelog.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Add _Checkstyle_ checks for _Java_ source files
+- Enable Quality Plugin by default for all Module plugins
+- Add _Checkstyle_ baseline generation using `./gradlew generateCheckstyleBaseline` task
+
 ## [0.3.0] - 2020-01-07
 
 ### Added